feat: 앱을 재실행해도 기존 예매 내역이 유지

+++ app/src/main/java/woowacourse/movie/dao/bookingStatus/BookingStatusDao.kt
@@ -11,7 +11,7 @@ interface BookingStatusDao {
     fun getAll(): List<BookingStatusEntity>
 
     @Insert
-    fun insert(bookingStatusEntity: BookingStatusEntity): Int
+    fun insert(bookingStatusEntity: BookingStatusEntity): Long
 
     @Delete
     fun delete(bookingStatusEntity: BookingStatusEntity)
